Object_detection_poc/app/requirements.txt
flask
ultralytics
pillow
torch 
numpy
gunicorn
matplotlib